TurnHeadingRoss
Created a new TurnHeading method to start from scratch and just get the
simplest functionality. Works well, as long as the magnetometer is
working. If the robot starts spinning, try resetting it. This is a
hardware issue I think I found a solution for, but Bahni would be a
better judge of that.