MATLAB code buff, Doc update, Parameter update
Added a uselessRun parameter to control dumping behavior in the case of
a useless run.

Also commented out a lot of unused functions. Consider deleting them...

Moved Lorenz code to the if(goingIn)-statement in the main loop, but I
want to talk with Will before I commit to this change - pretty confident
this is the way to go.

Removed checkPayload from goingOutModeRoss

Totally reworked the way that directions are checked in goingIn and
goingOut methods

Removed dumping_switch stuff