01/16 Week updates
moved lorenz/reversal stuff to robot selector so all trial stuff can be
set from the RobotSelector.h file,

Added a Charger testing method.

Wrote a determineState method to allwo the robots to find where they
were in the digging state machine prior to being turned on/reset.

Changed the minimum threshold needed for the MAG3110 sensor. Probably
need to tune this a bit more.

Added some stuff to wakepUp() which re-initializes the charging
detector. This resolved the reset on charger issue.

Got rid of the checkStart stuff in the fio code.

Added some followlane stuff to the visionmethods which allows the PID to
be called mroe frequently.