Lots of updates
I think it may be possible to remove the Fio from the robots altogether,
but in order to do this I need to demonstrate that the Due can perform
all of the tasks on the available pins.

I wrote the hardSerLCD library, so that the Due can now write to the LCD
screen on any of the open TX lines.

I also included the .pptx Vadim wrote for starting the robots in the
docs section.

I removed all of the fioWrite calls from the contact-action-handling
code.

I modified the CurrentSensor and Voltmeter sensor on the Fio to allow
better initialization on start up.