Lots of changes
Reorganized the codebase. Deprecated files from Mk2 are now in the
deprecated directory. New codebase is all together in due_digger_mk3
with all of the supporting code that is needed. Note that it currently
does not compile: still need to track down the relay and charger issues.
I think that it is just a pin assignment issue in either the def*.h or
the constants*.h files for A, but I could be wrong.